debug export env replace_envpl_var debug pwd and cd

+++ src/minishell.c
@@ -73,7 +73,7 @@ int minishell(t_solib *solib)
 	signal(SIGQUIT, &handle_signals);
 	g_signal_pid = 0;
 	pre_parsing(mini);
-	shlvl = soprintf_get(solib, "%d", ft_atoi(get_envpl_var(mini, "SHLVL")) + 1);
+	shlvl = soprintf_get(solib, "%d", (int)ft_atoi(get_envpl_var(mini, "SHLVL")) + 1);
 	replace_envpl_var(mini, "SHLVL=", shlvl);
 	rl_initialize();
 	while (mini->loop)
